feat(explorer): pagination and filters in url for parties
